+++ R/class-workbook-wrappers.R
@@ -482,43 +482,44 @@ wb_copy_cells <- function(
 #' @param wb A workbook object
 #' @param sheet A name or index of a worksheet
 #' @param dims worksheet cells
+#' @param solve logical if intersecting merges should be solved
 #' @param ... additional arguments
 #'
 #' @examples
 #' # Create a new workbook
-#' wb <- wb_workbook()
-#'
-#' # Add a worksheets
-#' wb$add_worksheet("Sheet 1")
-#' wb$add_worksheet("Sheet 2")
+#' wb <- wb_workbook()$add_worksheet()
 #'
 #' # Merge cells: Row 2 column C to F (3:6)
-#' wb <- wb_merge_cells(wb, "Sheet 1", cols = 2, rows = 3:6)
+#' wb <- wb_merge_cells(wb, dims = "C3:F6")
 #'
 #' # Merge cells:Rows 10 to 20 columns A to J (1:10)
-#' wb <- wb_merge_cells(wb, 1, cols = 1:10, rows = 10:20)
+#' wb <- wb_merge_cells(wb, dims = wb_dims(rows = 10:20, cols = 1:10))
 #'
-#' # Intersecting merges
-#' wb <- wb_merge_cells(wb, 2, cols = 1:10, rows = 1)
-#' wb <- wb_merge_cells(wb, 2, cols = 5:10, rows = 2)
-#' wb <- wb_merge_cells(wb, 2, cols = c(1, 10), rows = 12) # equivalent to 1:10
-#' try(wb_merge_cells(wb, 2, cols = 1, rows = c(1,10)))    # intersects existing merge
+#' wb$add_worksheet()
 #'
-#' # remove merged cells
-#' wb <- wb_unmerge_cells(wb, 2, cols = 1, rows = 1)  # removes any intersecting merges
-#' wb <- wb_merge_cells(wb, 2, cols = 1, rows = 1:10) # Now this works
+#' ## Intersecting merges
+#' wb <- wb_merge_cells(wb, dims = wb_dims(cols = 1:10, rows = 1))
+#' wb <- wb_merge_cells(wb, dims = wb_dims(cols = 5:10, rows = 2))
+#' wb <- wb_merge_cells(wb, dims = wb_dims(cols = 1:10, rows = 12))
+#' try(wb_merge_cells(wb, dims = "A1:A10"))
+#'
+#' ## remove merged cells
+#' # removes any intersecting merges
+#' wb <- wb_unmerge_cells(wb, dims = wb_dims(cols = 1, rows = 1))
+#' wb <- wb_merge_cells(wb, dims = "A1:A10")
+#'
+#' # or let us decide how to solve this
+#' wb <- wb_merge_cells(wb, dims = "A1:A10", solve = TRUE)
 #'
 #' @name wb_merge_cells
 #' @family workbook wrappers
 NULL
 
-
-
 #' @export
 #' @rdname wb_merge_cells
-wb_merge_cells <- function(wb, sheet = current_sheet(), dims = NULL, ...) {
+wb_merge_cells <- function(wb, sheet = current_sheet(), dims = NULL, solve = FALSE, ...) {
   assert_workbook(wb)
-  wb$clone()$merge_cells(sheet = sheet, dims = dims, ... = ...)
+  wb$clone(deep = TRUE)$merge_cells(sheet = sheet, dims = dims, solve = solve, ... = ...)
 }
 
 #' @export