+++ R/helper-functions.R
@@ -477,7 +477,7 @@ create_sparklines <- function(
       first = first,
       last = last,
       negative = negative,
-      "xr2:uid" = sprintf("{6F57B887-24F1-C14A-942C-%s}", random_string(length = 12, pattern = "[A-Z0-9]"))
+      "xr2:uid" = sprintf("{6F57B887-24F1-C14A-942C-%s}", random_string(length = 12, pattern = "[A-F0-9]"))
     ),
     xml_children = c(
       xml_node_create("x14:colorSeries", xml_attributes = colorSeries),
@@ -1212,4 +1212,127 @@ st_guid <- function() {
 #' @noRd
 st_userid <- function() {
   random_string(length = 16, pattern = "[a-z0-9]")
+}
+
+# solve merge helpers -----------------------------------------------------
+
+#' check side
+#' @param x a logical string
+#' @name sidehelper
+#' @noRd
+fullsided <- function(x) {
+  x[1] && x[length(x)]
+}
+
+#' @rdname sidehelper
+#' @noRd
+onesided <- function(x) {
+  (x[1] && !x[length(x)]) || (!x[1] && x[length(x)])
+}
+
+
+#' @rdname sidehelper
+#' @noRd
+twosided <- function(x) {
+  if (any(x)) (!x[1] && !x[length(x)])
+  else FALSE
+}
+
+#' @rdname sidehelper
+#' @noRd
+top_half <- function(x) {
+  if (twosided(x)) {
+    out <- rep(FALSE, length(x))
+    out[seq_len(which(x == TRUE)[1] - 1L)] <- TRUE
+    return(out)
+  } else {
+    stop("not twosided")
+  }
+}
+
+#' @rdname sidehelper
+#' @noRd
+bottom_half <- function(x) {
+  if (twosided(x)) {
+    out <- rep(TRUE, length(x))
+    out[seq_len(rev(which(x == TRUE))[1])] <- FALSE
+    return(out)
+  } else {
+    stop("not twosided")
+  }
+}
+
+#' merge solver. split exisisting merge into pieces
+#' @param have current merged cells
+#' @param want newly merged cells
+#' @noRd
+solve_merge <- function(have, want) {
+
+  got <- dims_to_dataframe(have, fill = TRUE)
+  new <- dims_to_dataframe(want, fill = TRUE)
+
+  cols_overlap <- colnames(got) %in% colnames(new)
+  rows_overlap <- rownames(got) %in% rownames(new)
+
+  # no overlap at all
+  if (!any(cols_overlap) || !any(rows_overlap)) {
+    return(have)
+  }
+
+  # return pieces of the old
+  pieces <- list()
+
+  # new overlaps old completely
+  if (fullsided(cols_overlap) && fullsided(rows_overlap)) {
+    return(NA_character_)
+  }
+
+  # all columns are overlapped onesided
+  if (fullsided(cols_overlap) && onesided(rows_overlap)) {
+    pieces[[1]] <- got[!rows_overlap, drop = FALSE]
+  }
+
+  # all columns are overlapped twosided
+  if (fullsided(cols_overlap) && twosided(rows_overlap)) {
+    pieces[[1]] <- got[top_half(rows_overlap), drop = FALSE]
+    pieces[[2]] <- got[bottom_half(rows_overlap), drop = FALSE]
+  }
+
+  # all rows are overlapped onesided
+  if (onesided(cols_overlap) && fullsided(rows_overlap)) {
+    pieces[[1]] <- got[, !cols_overlap, drop = FALSE]
+  }
+
+  # all rows are overlapped twosided
+  if (twosided(cols_overlap) && fullsided(rows_overlap)) {
+    pieces[[1]] <- got[, top_half(cols_overlap), drop = FALSE]
+    pieces[[2]] <- got[, bottom_half(cols_overlap), drop = FALSE]
+  }
+
+  # new is part of old
+  if (onesided(cols_overlap) && onesided(rows_overlap)) {
+    pieces[[1]] <- got[!rows_overlap, cols_overlap, drop = FALSE]
+    pieces[[2]] <- got[, !cols_overlap, drop = FALSE]
+  }
+
+  if (onesided(cols_overlap) && twosided(rows_overlap)) {
+    pieces[[1]] <- got[top_half(rows_overlap), cols_overlap, drop = FALSE]
+    pieces[[2]] <- got[bottom_half(rows_overlap), cols_overlap, drop = FALSE]
+    pieces[[3]] <- got[, !cols_overlap, drop = FALSE]
+  }
+
+  if (twosided(cols_overlap) && onesided(rows_overlap)) {
+    pieces[[1]] <- got[rows_overlap, top_half(cols_overlap), drop = FALSE]
+    pieces[[2]] <- got[rows_overlap, bottom_half(cols_overlap), drop = FALSE]
+    pieces[[3]] <- got[!rows_overlap, , drop = FALSE]
+  }
+
+  if (twosided(cols_overlap) && twosided(rows_overlap)) {
+    pieces[[1]] <- got[, top_half(cols_overlap), drop = FALSE]
+    pieces[[2]] <- got[, bottom_half(cols_overlap), drop = FALSE]
+    pieces[[3]] <- got[top_half(rows_overlap), cols_overlap, drop = FALSE]
+    pieces[[4]] <- got[bottom_half(rows_overlap), cols_overlap, drop = FALSE]
+  }
+
+  vapply(pieces, dataframe_to_dims, NA_character_)
 }